+++ src/components/elements/my-bottom-sheet/index.ts
@@ -1,0 +1,8 @@
+export { default } from './my-bottom-sheet'
+export type { MyBottomSheetProps, MyBottomSheetRef } from './type'
+export {
+  BottomSheetView,
+  BottomSheetScrollView,
+  BottomSheetFlatList,
+  BottomSheetTextInput,
+} from './bottom-sheet-primitives'